fix(Facility): Fixed locations being systematically deleted upon Facility update
OpenSILEX/opensilex-dev!1489

(cherry picked from commit 7a180546c8709e14c0cb4de59c198df82231c24e)

+++ opensilex-core/src/main/java/org/opensilex/core/location/bll/LocationObservationLogic.java
@@ -275,16 +275,20 @@ public class LocationObservationLogic {
 
     /**
      * Delete all location observations for a Collection
-     * Then deletes the collection to maintain coherence with objects that were never given a location (they have no collection)
+     * Then deletes the collection if doDeleteCollection is true to maintain coherence with objects that were never given a location (they have no collection)
      *
+     * @param session user client session
      * @param locationObservationCollectionURI location observation
+     * @param doDeleteCollection, if true then also deletes the containing collection
      */
-    public void deleteEveryLocationObservationInCollection(ClientSession session, URI locationObservationCollectionURI) throws Exception {
+    public void deleteEveryLocationObservationInCollection(ClientSession session, URI locationObservationCollectionURI, boolean doDeleteCollection) throws Exception {
         LocationObservationSearchFilter searchFilter = new LocationObservationSearchFilter();
         searchFilter.setObservationCollection(locationObservationCollectionURI);
 
         locationObservationDAO.deleteMany(session, searchFilter);
-        collectionLogic.deleteLocationObservationCollection(locationObservationCollectionURI);
+        if(doDeleteCollection){
+            collectionLogic.deleteLocationObservationCollection(locationObservationCollectionURI);
+        }
     }
 
     /**
